refactor(schema search): move SearchOutput to its own module
