+++ src/command/schema/search/mod.rs
@@ -1,15 +1,17 @@
+mod output;
+
 use std::{
     io::{self, Read},
     path::{Path, PathBuf},
 };
 
 use clap::Parser;
-use itertools::Itertools;
-use rover_schema::{ParsedSchema, SearchResult, root_paths::RootPath};
+use rover_schema::ParsedSchema;
 use rover_std::Fs;
 use serde::Serialize;
 
-use crate::{RoverOutput, RoverResult, command::CliOutput};
+use self::output::SearchOutput;
+use crate::{RoverOutput, RoverResult};
 
 #[derive(Debug, Serialize, Parser)]
 /// Search a GraphQL schema for types and fields by keyword
@@ -75,56 +77,4 @@ impl Search {
         let label = utf8_path.to_string();
         Ok((Fs::read_file(utf8_path)?, label))
     }
-}
-
-#[derive(Debug, Serialize)]
-pub struct SearchOutput {
-    query: String,
-    results: Vec<SearchResult>,
-}
-
-impl CliOutput for SearchOutput {
-    fn text(&self) -> String {
-        if self.results.is_empty() {
-            return format!("No results for \"{}\"", self.query);
-        }
-
-        let header = format!(
-            "{} result{} for \"{}\"",
-            self.results.len(),
-            if self.results.len() == 1 { "" } else { "s" },
-            self.query
-        );
-
-        let items = self.results.iter().map(format_result).join("\n\n");
-
-        format!("{header}\n\n{items}")
-    }
-
-    fn json(&self) -> Result<serde_json::Value, serde_json::Error> {
-        serde_json::to_value(self)
-    }
-}
-
-fn format_result(r: &SearchResult) -> String {
-    let first_line = match &r.description {
-        Some(desc) => format!("{} — {}", r.coordinate, desc),
-        None => r.coordinate.to_string(),
-    };
-
-    let kind_line = if r.via.is_empty() {
-        format!("  {}", r.kind)
-    } else {
-        let paths = r.via.iter().map(format_root_path).join(", ");
-        format!("  {}  ·  via {}", r.kind, paths)
-    };
-
-    format!("{first_line}\n{kind_line}")
-}
-
-fn format_root_path(p: &RootPath) -> String {
-    p.segments
-        .iter()
-        .map(|s| format!("{}.{}", s.type_name, s.field_name))
-        .join(" -> ")
 }
